Feature-cart

+++ jsp/removeProductProcess.jsp
@@ -2,13 +2,15 @@
 <%@include file="../db/connect.jsp"%>
 
 <%
+  int userId = (Integer)session.getAttribute("userId");
   int id = Integer.parseInt(request.getParameter("id"));
 
   if(id >= 0) {
     Connect con = Connect.getConnection();
 
-    String query = "DELETE FROM mycart WHERE FurnitureId = " + id;
+    String query = String.format("DELETE FROM mycart WHERE UserId = %d AND FurnitureId = %d", userId, id);
     con.executeUpdate(query);
+
     response.sendRedirect("productDetails.jsp");
   } else {
     response.sendRedirect("productDetails.jsp");
